requirements.txt
# Python dependencies for RSS Checker

# RSS feed parsing
feedparser>=6.0.10

# Discord integration
discord.py>=2.3.2
requests>=2.31.0

# Scheduling
schedule>=1.2.0

# Date/time handling
python-dateutil>=2.8.2

# AI features - Google Gemini API
google-generativeai>=0.3.0
